Move classes to their own directory
